+++ tests/test_manifest.js
@@ -64,6 +64,7 @@ export const schemeTestFiles = [
     'tests/core/scheme/time_tests.scm',
     'tests/core/scheme/eval_tests.scm',
     'tests/core/scheme/process_context_tests.scm',
+    'tests/core/scheme/bytevector_tests.scm',
 ];
 
 /**